Fix select filter count options that were not being updated on fragment refresh

+++ app/Filters/DataParser.php
@@ -95,7 +95,7 @@ class DataParser {
 	public function remove_filter( $id ) {
 
 		// Check if post is still available
-		if ( false !== get_post_status( $id ) ) {
+		if ( is_numeric( $id ) && false !== get_post_status( (int) $id ) ) {
 			wp_delete_post( $id, true );
 		}
 	}
